+++ tests/ops/integration_tests/saas/test_snap_task.py
@@ -5,15 +5,11 @@ from tests.ops.integration_tests.saas.connector_runner import ConnectorRunner
 from fides.config import CONFIG
 
 
-@pytest.mark.skip(reason="Currently unable to test OAuth2 connectors")
+# @pytest.mark.skip(reason="Currently unable to test OAuth2 connectors")
 @pytest.mark.integration_saas
-class TestsnapConnector:
+class TestSnapConnector:
     def test_connection(self, snap_runner: ConnectorRunner):
         snap_runner.test_connection()
-
-    # Run erasure with masking_strict = False so both update and delete actions can be used
-    masking_strict = CONFIG.execution.masking_strict
-    CONFIG.execution.masking_strict = False
 
     async def test_non_strict_erasure_request(
         self,
@@ -30,5 +26,4 @@ class TestsnapConnector:
             erasure_policy=erasure_policy_string_rewrite,
             identities={"email": snap_erasure_identity_email},
         )
-    # reset
-    CONFIG.execution.masking_strict = masking_strict
+        assert erasure_results == {"snap_instance:user": 6}